--- a/model/otlpgrpc/logs.go
+++ b/model/otlpgrpc/logs.go
@@ -162,6 +162,7 @@ func RegisterLogsServer(s *grpc.Server, srv LogsServer) {
 }
 
 type rawLogsServer struct {
+	otlpcollectorlog.UnimplementedLogsServiceServer
 	srv LogsServer
 }