--- a/model/otlpgrpc/metrics.go
+++ b/model/otlpgrpc/metrics.go
@@ -158,6 +158,7 @@ func RegisterMetricsServer(s *grpc.Server, srv MetricsServer) {
 }
 
 type rawMetricsServer struct {
+	otlpcollectormetrics.UnimplementedMetricsServiceServer
 	srv MetricsServer
 }