--- a/model/pdata/spanid.go
+++ b/model/pdata/spanid.go
@@ -29,6 +29,9 @@ func InvalidSpanID() SpanID {
 }
 
 func NewSpanIDFromBytes(bytes []byte) SpanID {
+	if len(bytes) != 8 {
+		return SpanID{orig: data.NewSpanID([8]byte{0})}
+	}
 	var b [8]byte
 	copy(b[:], bytes[:8])
 	return SpanID{orig: data.NewSpanID(b)}