--- a/model/pdata/traceid.go
+++ b/model/pdata/traceid.go
@@ -35,7 +35,7 @@ func NewTraceID(bytes [16]byte) TraceID {
 
 func NewTraceIDFromBytes(bytes []byte) TraceID {
 	if len(bytes) != 16 {
-		return TraceID{orig: data.NewTraceID([16]byte{0})}
+		return TraceID{}
 	}
 	var b [16]byte
 	copy(b[:], bytes[:16])