… internal/testdata/trace.go
… model/internal/data/protogen/collector/logs/v1/logs_service.pb.go
… model/internal/data/protogen/collector/logs/v1/logs_service_vtproto.pb.go
… model/internal/data/protogen/collector/metrics/v1/metrics_service.pb.go
… model/internal/data/protogen/collector/metrics/v1/metrics_service_vtproto.pb.go
… model/internal/data/protogen/collector/trace/v1/trace_config.pb.go
… model/internal/data/protogen/collector/trace/v1/trace_config_vtproto.pb.go
… model/internal/data/protogen/collector/trace/v1/trace_service.pb.go
… model/internal/data/protogen/collector/trace/v1/trace_service_vtproto.pb.go
… model/internal/data/protogen/common/v1/common.pb.go
… model/internal/data/protogen/common/v1/common_vtproto.pb.go
… model/internal/data/protogen/logs/v1/logs.pb.go
… model/internal/data/protogen/logs/v1/logs_vtproto.pb.go
… model/internal/data/protogen/metrics/v1/metrics.pb.go
… model/internal/data/protogen/metrics/v1/metrics_vtproto.pb.go
… model/internal/data/protogen/resource/v1/resource.pb.go
… model/internal/data/protogen/resource/v1/resource_vtproto.pb.go
… model/internal/data/protogen/trace/v1/trace.pb.go
… model/internal/data/protogen/trace/v1/trace_vtproto.pb.go
… model/internal/otlp_wrappers_test.go
… model/pdata/common_test.go
… model/pdata/generated_metrics.go
… model/pdata/generated_metrics_test.go
… model/pdata/generated_trace.go
… model/pdata/metrics_test.go
… proto_patch.sed