**Description:** <Describe what has changed. 
This is attempt number 3 to address the issue of supporting optionals in the proto. This approach removes the use of gogo completely, and uses the google protobuf library in combination with the vtproto extension.

NOTE: This is still very much in progress, early benchmarks are not encouraging.

Before the change:
```
pkg: go.opentelemetry.io/collector/model/pdata
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkAttributeValue_CopyTo-12                                	455596772	         3.337 ns/op
BenchmarkAttributeValue_SetIntVal-12                             	68716096	        15.82 ns/op
BenchmarkAttributeMap_Range-12                                   	22402864	        46.87 ns/op
BenchmarkAttributeMap_RangeOverMap-12                            	 6150018	       193.4 ns/op
BenchmarkStringMap_RangeOverMap-12                               	 5959550	       196.1 ns/op
BenchmarkLogsClone-12                                            	   38168	     32231 ns/op
BenchmarkMetricsClone-12                                         	     576	   2099450 ns/op
BenchmarkOtlpToFromInternal_PassThrough-12                       	15396999	        83.52 ns/op
BenchmarkOtlpToFromInternal_Gauge_MutateOneLabel-12              	15304377	        77.29 ns/op
BenchmarkOtlpToFromInternal_Sum_MutateOneLabel-12                	15054184	        67.47 ns/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-12    	20351544	        57.17 ns/op
BenchmarkTracesClone-12                                          	    2844	    425877 ns/op
```

After the change:
```
pkg: go.opentelemetry.io/collector/model/pdata
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkAttributeValue_CopyTo-12                                	348510321	         3.405 ns/op
BenchmarkAttributeValue_SetIntVal-12                             	97121374	        12.32 ns/op
BenchmarkAttributeMap_Range-12                                   	27596403	        42.68 ns/op
BenchmarkAttributeMap_RangeOverMap-12                            	 6191503	       191.8 ns/op
BenchmarkStringMap_RangeOverMap-12                               	 5902920	       197.4 ns/op
BenchmarkLogsClone-12                                            	   12105	    101604 ns/op
BenchmarkMetricsClone-12                                         	     214	   4715739 ns/op
BenchmarkOtlpToFromInternal_PassThrough-12                       	18306332	        62.48 ns/op
BenchmarkOtlpToFromInternal_Gauge_MutateOneLabel-12              	20573402	        57.43 ns/op
BenchmarkOtlpToFromInternal_Sum_MutateOneLabel-12                	19648472	        60.16 ns/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-12    	19755382	        59.61 ns/op
BenchmarkTracesClone-12                                          	    1064	   1126385 ns/op
```

**Link to tracking Issue:** #4258